Update FirstSteps.pod from Mike Burns input
Standard indent for interactive examples is 2 spaces
and leave the pdl> prompt in the text to indicate
which is command and which is interactive.  Also,
be sure to include ; on any command that could be
used in a program file to avoid user annoyances when
a simple cut-n-paste doesn't work.