Updated PP chapter with threadloop gotchas and /* comments */
With commit 757691190bb503aea6e35a98106601, PDL no longer chokes
when users try to comment-out a piece of code with a memory access
or loop(). I have revised the text to be aware of this change.

Also, in the process of writing the tests for that change in PDL,
I discovered a gotcha with threadloop blocks. I've added a few
paragraphs to discuss those issues and how to avoid them.